LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/back-forward-out-of-bounds-expected.txt



FAIL back() and forward() out of bounds assert_equals: committed and finished must reject with the same value expected object "InvalidStateError: Cannot go back" but got object "InvalidStateError: Cannot go back"
